css fixes, unified profile
