--- frontend/src/pages/AdminProfile/index.css
@@ -1,40 +1,0 @@
-.admin-profile-container {
-    display: flex;
-    flex-direction: column;
-    padding: 20px;
-    font-family: var(--font-secondary);
-    width: 100%;
-    box-sizing: border-box;
-}
-
-.admin-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 20px;
-}
-
-.admin-title {
-    flex: 1;
-    color: var(--primary-blue);
-    text-decoration: none;
-    font-family: var(--font-secondary);
-    font-style: normal;
-    font-weight: 700;
-    font-size: var(--page-title-text);
-  }
-
-.logout-button {
-    height: fit-content;
-    padding: 0.6vw 0.85vw;
-    background-color: var(--white);
-    border: 1px solid var(--primary-blue);
-    border-radius: 8px;
-    box-shadow: 0px 4px 16px #0f111105;
-    cursor: pointer;
-    font-family: var(--font-primary);
-}
-
-.logout-button i {
-    color: var(--dark-grey);
-}
