--- frontend/src/pages/AdminVerify - ToDelete/index.css
@@ -1,41 +1,0 @@
-.table-data {
-    padding-top: 1vw;
-}
-
-.verify-table {
-    width: 80vw;
-    margin: auto;
-    border-collapse: collapse;
-    border-spacing: 0;
-    font-family: var(--font-secondary);
-}   
-
-.verify-table-head .verify-th {
-    color: black;
-    height: 2vw;
-    font-family: var(--font-secondary);
-}
-
-.verify-table .verify-tr:hover {
-    background-color: #f5f5f5;
-}
-
-.verify-table .verify-th, .verify-td {
-    border: 1px solid #ddd;
-    text-align: left;
-    padding: 0.7vw;
-}
-
-.verify-table .verify-td {
-    font-family: var(--font-tertiary);
-}
-
-.errorMsg {
-    text-align: center;
-}
-
-.verify-heading {
-    padding-top: 2vw;
-    padding-left: 10vw;
-    font-family: var(--font-secondary);
-}
