+++ frontend/src/pages/Profile/index.css
@@ -3,9 +3,9 @@
 .content {
     display: flex;
     font-family: var(--font-secondary);
-    margin-bottom: 4vh;
-    padding: 0 36px;
     gap: 20px;
+    padding: 0 36px 36px;
+    flex-grow: 1;
 }
 
 .column-1,
