+++ frontend/src/pages/Settings/index.css
@@ -1,7 +1,7 @@
 .settings-container {
     display: flex;
     flex-direction: column;
-    padding: 20px;
+    padding-inline: 36px;
     font-family: var(--font-secondary);
     width: 100%;
     box-sizing: border-box;
